requirements.txt
torch==0.3.0.post4
Pillow==5.0.0
torchvision==0.2.0
bcolz==1.1.2
cloudpickle==0.2.2
Pillow==5.0.0
matplotlib==2.0.2
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
numpy==1.12.1
pandas==0.20.1
kaggle-cli==0.12.13